Restructure platform specific code
We intend to have a single factory function for all platforms. This allows us to add some assertions without having to worry about inheritance.
